imagine alb-negru

+++ src/backend/opencl/pixel.cl
@@ -1,6 +1,6 @@
 __kernel void rgbToBw(__global unsigned char* in, __global unsigned char* out)
 {
-    int bwIndex = get_global_id(0);
-	int rgbIndex = bwIndex * 3;
+    unsigned long bwIndex = get_global_id(0);
+	unsigned long rgbIndex = bwIndex * 3;
 	out[bwIndex] = (in[rgbIndex] + in[rgbIndex+1] + in[rgbIndex+2]) / 3;
 }
